packages/seeding/requirements.txt
google-cloud-aiplatform[generative_ai]
google-cloud-storage
